Add exceptions handling, string formatting and database scripts

+++ waskoDB_script.sql
@@ -5,8 +5,7 @@ CREATE TABLE cars (
 	id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
 	manufacturer CHAR(15) NOT NULL,
 	model CHAR(15) NOT NULL,
-	capacity FLOAT NOT NULL
+	capacity DOUBLE(5,3) NOT NULL
 )
 ENGINE = innodb
 DEFAULT CHARACTER SET utf8 COLLATE utf8_unicode_ci;
-
